+++ sdk/lib/_internal/js_dev_runtime/private/ddc_runtime/errors.dart
@@ -90,7 +90,10 @@ throwNullValueError() {
   // TODO(vsm): Per spec, we should throw an NSM here.  Technically, we ought
   // to thread through method info, but that uglifies the code and can't
   // actually be queried ... it only affects how the error is printed.
-  throw NoSuchMethodError(null, Symbol('<Unexpected Null Value>'), null, null);
+  throw NoSuchMethodError.withInvocation(
+    null,
+    Invocation.method(Symbol('<Unexpected Null Value>'), null, null),
+  );
 }
 
 castError(obj, expectedType) {